services/clip_nsfw/requirements.txt
# Requirements placeholder

torch>=2.1.0
transformers>=4.35.0
Pillow>=10.1.0
numpy>=1.24.0
opencv-python>=4.8.0
python-multipart>=0.0.6
pydantic>=2.4.2
fastapi>=0.104.0
uvicorn>=0.24.0
